import schedule screen
