+++ src/app/(user)/sign-out.tsx
@@ -63,6 +63,7 @@ const styles = StyleSheet.create({
     shadowOffset: {width: 0, height: 4},
     shadowRadius: 4,
     shadowOpacity: 1,
+    elevation: 10,
   },
 
   noButton: {
