+++ src/modeus/parser.ts
@@ -1,4 +1,4 @@
-import { ScheduleData } from '@/app/(user)';
+import { ScheduleData } from '@/app/(user)/schedule';
 import moment from 'moment';
 import 'moment/locale/ru';
 export const parseICS = (icsString: string) => {
